Storage skip: change in storage skipping
In testing ovirt 4.2 we change approach how to generate tests for particular
storage types. Now we are using parametrize of fixture called storage, but
problem is that this fixture is parametrized after pytest_runtest_setup.
Pytest put all parametrized values of test to item._genid which is e.g.
string "nfs-foo-bar" in case you parametrized that test with nfs, foo,
bar.

This change is backward compatibility because I it's just looking for used
fixture storage and if it's there it's using new approach otherwise using old
approach.

Change-Id: Ie0bb214cad2251b5a236446307df3cc145ad4bd4